Changed a few more front end things, like how speakers show up in the main table, and showing speakers in all the different views.
Made it so the List<string> Speakers gets converted to json and stored in the database as a json string, then it gets rehydrated to be displayed by the different views. This is done in the ProgramContext.cs

Because the List<string> cannot have a RegularExpression data annotation on it, I went ahead and built my own custom validation attribute.